group cloud compliance results by control ids
